@@ -1,19 +1,19 @@
 module github.com/prometheus/client_golang
 
-go 1.21
+go 1.22
 
 require (
 	github.com/beorn7/perks v1.0.1
 	github.com/cespare/xxhash/v2 v2.3.0
-	github.com/google/go-cmp v0.6.0
+	github.com/google/go-cmp v0.7.0
 	github.com/json-iterator/go v1.1.12
-	github.com/klauspost/compress v1.17.11
+	github.com/klauspost/compress v1.18.0
 	github.com/kylelemons/godebug v1.1.0
 	github.com/prometheus/client_model v0.6.1
-	github.com/prometheus/common v0.61.0
-	github.com/prometheus/procfs v0.15.1
-	golang.org/x/sys v0.28.0
-	google.golang.org/protobuf v1.36.1
+	github.com/prometheus/common v0.63.0
+	github.com/prometheus/procfs v0.16.0
+	golang.org/x/sys v0.30.0
+	google.golang.org/protobuf v1.36.6
 )
 
 require (
@@ -23,9 +23,9 @@ require (
 	github.com/modern-go/reflect2 v1.0.2 // indirect
 	github.com/munnerz/goautoneg v0.0.0-20191010083416-a7dc8b61c822 // indirect
 	github.com/mwitkow/go-conntrack v0.0.0-20190716064945-2f068394615f // indirect
-	golang.org/x/net v0.32.0 // indirect
-	golang.org/x/oauth2 v0.24.0 // indirect
-	golang.org/x/text v0.21.0 // indirect
+	golang.org/x/net v0.35.0 // indirect
+	golang.org/x/oauth2 v0.25.0 // indirect
+	golang.org/x/text v0.22.0 // indirect
 	gopkg.in/yaml.v2 v2.4.0 // indirect
 )
 
